Moved context types to d.ts file
