--- src/context/web3/IWeb3.ts
@@ -1,10 +1,0 @@
-import { Provider } from "ethers/providers";
-
-export interface Action {
-  payload?: any;
-  type: string;
-}
-
-export interface Web3State {
-  provider: Provider;
-}
